tapdb/postgres: attempt to catch serialization error to retry
In this commit, we introduce a new postgres specific error returned when
a transaction cannot be serialized. We then catch this error in order to
implement retry logic for a transaction that failed to be serialized.
We'll retry up to 10 times, waiting a random amount up to 50 ms between
each retry.