tapfreighter: resume pending parcels using the main porter goroutine
In this commit, we ensure that pending parcels are resumed using the
same goroutine that processes the normal send parcel processing flow.

The goal of this change is to ensure that the locking mechanism enforced
by a singular parcel processing goroutine is not compromised by the
execution of a secondary parcel goroutine.