tapcfg: ensure log rotator created after config validation
In this commit, we fix a bug that would cause the logs to always go to
the default log location even if a different base dir was set. Before
this commit, we'd create the log rotators before the call to
`ValidateConfig`, which is where we modify the log dir location to be
relative to the final base dir.

The fix here is simple: create the log rotator _after_ the config has
been validated.