tapd: detect SQLITE_BUSY and re-try
If SQLITE_BUSY is returned, it means waiting for a transaction failed
because too many other readers/writers are currently using the DB. We
can detect that error, convert it into a serialization error and detect
that in the existing retry loop.
We also re-try if creating or committing a transaction fails with a
serialization error.