tapcfg+tapdb: use clock for universe DB
In a next commit we'll want to be able to better test asset events which
are always inserted into the DB with a timestamp of `time.Now()`.
In order to be able to influence what time is actually written to the
DB, we use the `clock.Now()` value that can be overwritten in unit
tests.