tapdb+asset: handle group sigs as a witness stack
In this commit, we update the DB logic to use the new WitnessStack field
for queries related to asset groups. We add helpers to convert between a
typed TxWitness and a byte array, and restict the witness handling logic
to only allow a single signature to maintain the same functionality.