monitoring: add new prometheus monitoring pkg
Wrap prometheus calls in the `PrometheusExporter` struct. The
`PrometheusExporter` will be used in the application code to set custom
metrics.

The prometheus server will not run by default. The project needs to be
build with the `monitoring` tag to include the exporter's code and
enabled in the config to run.