multi: automate updating of proof test files/vectors
There were some test and test vector files that weren't yet
automatically updated when running "make gen-test-vectors". To allow
easy updating of any proof related test files in case anything in the
encoding changes, we modify an existing integration test to generate new
contents for these files.

Because an integration test run is not deterministic as we always spin
up a new regtest chain, these files cannot be included in the
test-vector-check make goal that is run on the CI.
Instead the "make gen-test-vectors" goal should be called manually
whenever there is a change to the proof encoding format to trigger an
update of these files.