multi: add new virtual output type for burns
Because a burn is an interactive send, we don't need to create a
tombstone output if it's a full-value send. So in case there are passive
assets in the same commitment, we don't have a split root output to hook
them to. With the introduction of the new TypeSimplePassiveAssets we
have a new way to signal to the send state machine where to commit
passive assets to if there is no actual split or tombstone output.