tapdb: materialize burned assets as spent
We want to be able to show/list burned assets, so we want to re-create
them in the assets table. But to avoid them showing up in any balances,
we also mark burned assets as spent and we'll add an extra is_burn flag
to the asset output as well on the RPC response.