tapgarden: give synchronous feedback for batch finalization
Before, the batch finalization would always happen in the background and
only the state of the batch before freezing it would be returned.
With this change we actually wait for the batch transaction to be
created and broadcast before returning to the caller. This allows us to
return an updated state of the batch or an error in case one occurs.
This also prevents us from setting the pending batch to nil even if it
fails to be created (for example if there are no funds to create the
mint transaction). So a finalizatino can be re-attempted in case of a
temporary error.