itest: use correct testing struct for t.Cleanup()
It turns out that we used ht.t for the postgres harness, so the
ht.t.Cleanup() would only be triggered at the very end of the whole
integration test run and not after each individual test case.
This lead to many Postgres containers being created which the GitHub CI
didn't like (it killed our process).
By passing in the sub-test specific testing struct, we avoid this issue.