CLI: Improve minting user experience (#492)
* multi: return full batch on mint and finalize

* tapgarden: give synchronous feedback for batch finalization

Before, the batch finalization would always happen in the background and
only the state of the batch before freezing it would be returned.
With this change we actually wait for the batch transaction to be
created and broadcast before returning to the caller. This allows us to
return an updated state of the batch or an error in case one occurs.
This also prevents us from setting the pending batch to nil even if it
fails to be created (for example if there are no funds to create the
mint transaction). So a finalizatino can be re-attempted in case of a
temporary error.

* itest: fix batch state validation

* tapdb/sqlc: add asset_id to QueryUniverseLeaves query

* multi: add grouped assets to universe roots call

* taprpc: rename Batch to PendingBatch for Mint call

* taprpc+tapgarden: obtain pending batch directly from response chan

We can skip the on chain disk hit and instead can just rely on the
planter to send us the batch directly.

---------

Co-authored-by: Olaoluwa Osuntokun <laolu32@gmail.com>