Provable asset burning (#477)
      * asset+make: add DeriveBurnKey function

* asset: add IsBurn method

* multi: add new virtual output type for burns

Because a burn is an interactive send, we don't need to create a
tombstone output if it's a full-value send. So in case there are passive
assets in the same commitment, we don't have a split root output to hook
them to. With the introduction of the new TypeSimplePassiveAssets we
have a new way to signal to the send state machine where to commit
passive assets to if there is no actual split or tombstone output.

* tapfreighter: add FundBurn method to wallet

* tapdb: materialize burned assets as spent

We want to be able to show/list burned assets, so we want to re-create
them in the assets table. But to avoid them showing up in any balances,
we also mark burned assets as spent and we'll add an extra is_burn flag
to the asset output as well on the RPC response.

* tapfreighter: prune burned assets from commitment

The same way we prune tombstone asset outputs from the input commitment
when spending an anchor output, we also want to remove any burned
assets.

* taprpc+tapfreighter: add BurnAsset RPC method

* cmd/tapcli: add assets burn sub command

* tapscript: allow collectible full-value send with passive assets

In the case where we have an interactive full-value send of a
collectible, we would normally only expect a single output. But if there
are passive assets to commit to, then there needs to be a second output
just for those.

* itest: add burn tests