asset: introduce new asset version v1 for segwit-like encoding
In this commit, we add a new asset version to introduce segwit-like
asset encoding where we don't encode the witness vector of an asset's
inputs. This applies on the asset version level, so all v1 asset won't
have that vector encoded. Note that this is _only_ for the MS-SMT tree,
any other TLV encoding such as the asset proof or state transitions
remain the same.

Fixes https://github.com/lightninglabs/taproot-assets/issues/464