universe: update pushProofToFederation to make sure all servers are tried
In this commit, we fix a bug that would cause universe proof push to
fail for all servers, if anyone of them was unreachable, or rejected the
proof.

If we return `err` here, then eventually the context will be read by
other goroutines, which is already cancelled at this point. This'll
cause the push attempt to stop short.

We now long and return `nil` so the other goroutines will continue.

See https://github.com/lightninglabs/taproot-assets/issues/527 for more
details.