tapgarden+tapdb: store group anchors early
In this commit, we address the issue of asset insertion ordering during
minting. Group anchors must be inserted before reissued assets in the
same batch for group info to be stored correctly and asset verification
to succeed. We also update configs for other minting subsystems to
propogate group verifier callbacks.