multi: update split locators to include asset version
In this commit, we update split locator struct to include the asset
version. This ensures that if we attempt a send with heterogeneous asset
versions, then each vOut is able to ensure that the split is created
with the correct asset version.

Without this, we might use a different asset version with the proof we
have on disk vs what we put in the transaction.

We also modify the way the change output is created. The change output
will now always take on the max version of all the inputs.