loadtest: allow test case to be specified on command line too
Because we are running a compiled Go test case, we can't use any custom
flags as the internal flag parser only accepts the pre-defined
"-test.XXX" flags. But we couldn't use those either because we attempted
to also parse the flags with the lnd flag parsing library.

This commit fixes the issue by removing any custom flag parsing and
allowing test cases to be specified using the -test.run flag.
For example:

	-test.run="^TestPerformance/mint_batch_stress$"