tapscript: rename variables for clarity, add trace logs
This is an area that is frequently visited when debugging commitment
based problems.
To improve clarity on some of the variable names, we perform a series of
renames. We also add in some trace level log statments.

This commit does NOT change any behavior in the code (other than
logging).