tapfreighter: remove tombstones and burns from commitment
This fixes a bug that we only discovered after sending assets many times
back and forth. We didn't properly remove tombstones and burns from the
commitments when creating new outputs. We did have tests that sent
passive assets out from a commitment that had a tombstone. But we never
spent that second output, which would have discovered the bug earlier.
We'll fix this with a new integration test case in the next commit.