rpc: unconstrain meta field for MintAssets
In this commit, we stop trying to strictly parse the meta field.
Instead, users can set it to w/e they want, with the default being
opaque. This lets users start to add structure to the meta field from
day one. One example is a set meta with fields that dictate how the
asset unit is to be displayed. This can be used to do things like
leverage the current fix point format to add decimal places for display.