tapdb: fix overwriting of default configs
In this commit, we fix an issue with the way we returned the global
configs. We always assumed that there would be two items in the db.
However if we only updated one of them, then only a single version will
be returned. This means we'll only return one config, when we should
always return two.