tapdb: store namespace with universe sync config
In this commit, we fix an issue with universe sync config storage. With
the old table schema, multiples entries could exist for the same
Universe because the UNIQUE constraint did not function as intended.
This was caused by NULL entries being treated as unique values vs. one
value. The new table has one entry per Universe ID, which fixes this.