universe: sort incoming transfer proofs by block height
In this commit, we attempt a simple sorting routine to address:
https://github.com/lightninglabs/taproot-assets/issues/591.

Without these, we'll try to validate transfer proofs out of order, which
is akin verifying an orphan transaction in the Bitcoin mempool. Rather
than do a full on topological sort, we instead just sort by block
height. As we don't support spending unconfirmed change, this should be
enough to serialize the contents of a block.