tapscript: fix wrong index being used
This fixes an error where index 0 is not equal to the recipient's index.
This is only the case for burns so it wasn't detected before.
For a burn, we always have two outputs: One for the split root and one
for the actual burn. If a split root isn't needed because it's a burn of
the full amount (and we're just carrying passive assets), we actually
remove the first output again. But that happens _after_ this code, so we
need to make sure we use the correct output.