tapdb: reset split commitment for passive assets
Once an asset from a split becomes a passive asset in the next transfer,
we can actually drop the split commitment. We did that in the commitment
but not in the database, which resulted in the split commitment being
present again in the input commitment (which lead to a root key
mismatch).
This wasn't discovered before because we didn't have a test that carried
an asset resulting from a split as a passive asset in the next transfer.