universe: syncer defaults to syncing universes with specified configs
Previously the syncer defaulted to syncing all universes (global sync).
With this commit, the universe only performs a global sync if the global
sync config is enabled for insertion. Notably this skips retrieving all
universe roots in the default case.

The default is now to sync only those universes which have enabled
universe specific configs.