address+universe: add AssetSyncer interface
In this commit, we add an interface to enable the address Book to create
addresses for assets the node has not bootstrapped yet. This includes
assets that are part of a group, but the node did not yet import the
asset group itself. The node will try to fetch a matching Universe root
from all universe servers in its federation.