tapdb: ensure universe asset stats ticker is stopped before refresh
In this commit, we fix a bug that would cause the universe stats cache
to be wiped more than once every 30 minutes. Before this commit, we'd
accumulate a new background `time.AfterFunc` ticker for each unique
request set in the cache.

To fix this, we'll always first stop the old ticker before making a new
one.