universe: allow FetchMultiverseLeaves to multi-query
In this commit, we modify the FetchMultiverseLeaves method to allow it
to return either all the leaves, or a subset of leaves filtered by the
`MultiverseLeafDesc` type. In the future, we can further extend this to
use postgres batch queries so we can get the resp in a single round
trip.