taprpc: add newGroupedAsset and GroupedAsset flags
In this commit, we replace the enableEmission flag for minting with two
flags, newGroupedAsset and GroupedAsset. These are mutually exclusive
and less ambigious than the parsing of enableEmission. newGroupedAsset
is used when creating a group anchor, and groupedAsset is used when
issuing into an existing group. We also add the MintSeedling type to
better display the state of a pending batch, as the enableEmission
variable was missing before.