tapgarden: remove extra error sent during cancel
In this commit, we remove an extra error broadcast during caretaker
cancellation that could prevent graceful shutdown. If the caretaker
state machine has not reached BatchStateBroadcast, it sends potential
errors to the planter on a channel with capacity of one. If cancellation
is requested before reaching BatchStateBroadcast and fails internally,
sending that error to the planter prevents an error from being sent by
the main caretaker goroutine. We also unify cancel request handling.