tapdb: add test that demoes DB migration testing
This test is mostly a demo for how to use the migration testing
framework.
We start with the DB that only contains the very first migration file,
which we assert by making sure we encounter a schema error when
attempting to read assets.
We then move forward to version 11, insert some dummy data, then run all
migrations up to the latest version.