tapgarden: remove duplicate batch freeze
In this commit, we update the caretaker start logic to remove an
unnecessary batch write. Before we start the caretaker, we write the
batch with the Frozen state, but we don't update the in-memory pending
batch to move from the Pending to Frozen state. This causes the
caretaker to write the batch again on start. We can address this by
updating the in-memory batch after a successful batch freeze.