tapgarden: stop caretaker on confirmation err
In this commit, we update the TX confirmation handling logic to stop the
caretaker if confirmation registration fails. At that point, the
caretaker cannot successfully receive a confirmation for the broadcast
batch, so it should shut down to allow caretaker restart for the same
batch. Note that the planter will not actually delete the stopped
caretaker, as the error is not sent on BroadcastErrChan.