universe: fix TCP connection leak by explicitly closing gRPC conn for syncers
In this commit, we attempt to fix a TCP connection leak by explicitly
closing the gRPC connections we create once we're done with the relevant
gRPC client. Otherwise, we'll end up making a new connection for each
new asset to be pushed, which can add up. In the future, we should also
look into the server-side keep alive options.