gitignore: add non-debug binaries to ignore
During the release build (when not running with SKIP_VERSION_CHECK=1),
tapd is built in the local directory to extract the version.
Because tapd wasn't in the .gitignore file before, that caused a
manifest mismatch on the built release binaries, because the attribute
build.vcs.modified=true was compiled into the binaries.