proof: extract FetchProofProvenance function for iterative retrieval
With this commit we extract the logic for iteratively retrieving the
full provenance for an asset by starting at the last proof then querying
each previous proof until we arrive at the genesis.
We will want to re-use this logic outside of the proof courier itself,
so it's useful to extract into a general-purpose function where the
actual retrieval of an individual proof is done in a callback and can be
adjusted to the current proof storage (local or remote).