multi: add HasProof to proof.Archive interface
In some situations we want a quick way to find out if we have a proof
(or not) without actually fetching it, so we add a HasProof method to
the proof.Archive interface.
The MultiArchiver will return false if one of the backends does not have
the proof, so we can use that method to find out if we have a
discrepancy between the backends and should import a proof file.