tapgarden: refactor Custodian.receiveProof so that it's sync agnostic
This commit refactors the method `Custodian.receiveProof` such that it's
sync/goroutine agnostic (can be run with/without a goroutine). This
change moves all goroutine wait group management code to the point where
the goroutine is started.

This commit modifies `receiveProof` such that it returns an error rather
than writes an error directly to log. This means that we can pass the
error to the custodian's error channel.