rpcserver+tapscript: pre-fill and use key info from vPSBT
When signing a virtual PSBT, we don't want to use the key descriptor
info of the input asset's script key, as that information is not
available when the input asset is created from a de-serialized asset.
So previous to the change in this commit the SignVirtualPsbt RPC method
could only be called on vPSBTs that were funded through FundVirtualPsbt
that populated the key information.
After this commit it is possible to sign a vPSBT that was created from
an input proof only since the required key information is loaded from
the database before signing.