asset: update DeepEqual to work with asset v1 (segwit)
The asset version v1 (segwit) allows an asset to be committed without
the TxWitness field. Therefore, when comparing two such assets, we must
also ignore the TxWitness field.

If we don't do this, then the deep equal check when creating a proof for
a split asset fails when checking if the split root asset was committed
to the root tree.