tapdb: make sure proofs don't collide
With this commit we make sure that proofs in the database don't collide
if they use the same script key but different anchor transactions.
This situation can occur if a TAP address receives multiple transfers.
We also make sure we can query for the correct proof if we also specify
the outpoint in the query.