tapdb: fix proof update by ID
Since we now have a JOIN with the managed_utxos in the UpsertAssetProof
we cannot use that to update proofs that haven't been re-anchored yet.
But because we re-anchor passive assets only _after_ updating the
proofs, this would previously result in no rows being updated.
We could change the order of operations instead, but having WHERE clause
for a specific database ID mixed in with optional value based queries
wasn't super beautiful in the first place. So we opt for having a more
explicit upsert for database ID based proof identification.