proof: use outpoint in proof file name
To make sure we don't accidentally overwrite a proof file if we receive
to a TAP address multiple times, we also use the first couple of bytes
of the outpoint TXID and index in the file name.
We don't use the full outpoint as in some operating systems the full
path for a file is not allowed to exceed 256 characters (the path and
file name combined). And since we already use 130 characters for the hex
encoded asset ID and script key, we need to shorten the outpoint
somewhat.
We will add a migration that renames existing files on disk in the next
commit.