tapdb+tapfreighter: don't use locator hash as map key
To get a unique lookup key we mis-used a proof locator's hash as the key
to look up passive asset proofs quickly without needing to scan a slice.
Because we don't have the same data available when creating the map as
we do when accessing it further down in the asset database, we got a
mismatch in the hash and proofs couldn't be found (mainly due to not
having access to the previous outpoint of the passive assets being
re-anchored).
As a compromise we now map the proofs by asset ID and have a slice of
proofs in case there are multiple passive assets with the same asset ID.
This is better than having a completely flat slice of proofs (as we
don't have to scan through all of them) but still requires us to do
_some_ iterating.