tapdb: turn InsertAssetWitness into UpsertAssetWitness
Because we can now end up importing proofs that we already have (for
example when sending to our own TAP address using a universe courier, we
will pull the proof from the local universe and import it into our store
again).
Before turning this into an upsert, we would end up with an asset that
had two identical entries in the previous witness list.
To make sure the order of multiple witnesses is kept, we also need to
add a witness_index field that we can use for sorting.
Unfortunately that breaks the data migration demo test as the
FetchAllAssets query also loads the witness and the query for that uses
the witness_index field that doesn't exist at that point.
Since the current test is only a demo we just change it to fetch the
(unchanged) managed UTXOs instead.