tapfreighter: fix wrong proof outpoint being used
This fixes a couple of instances where the porter used the wrong
outpoint for proof locators when creating new proofs for a send package.
This was previously not noticed because the outpoint in the locator was
ignored by both the file based and database archive. With the outpoint
now being mandatory, this lead to failures in the integration tests.