tapgarden: fix proof not found bug
This fixes a bug reported by a user running v0.3.3-rc1. Although the
situation can only happen if the daemon is shut down in exactly the
wrong moment (or, more likely, due to an otherwise inconsistent database
state), it can happen that the multiverse reports a proof is available
but it wasn't yet imported into the local archive.
To make sure we can definitely rely on the proof being in the asset DB
when trying to complete a receive event, we double check and re-import
the proof if necessary.