tapdb: add new tapscript tree tables and queries
This commit adds three new tables that enable us to store tapscript
trees, with deduplication of tapscript leaves. A tree can also be stored
as two TapHashes representing a TapBranch, allowing for externally
managed tapscript trees. These tables are not intended to be modified
with the basic queries added in this commit, but wrapper functions in
the next commit.