itest: use configurable timeout for load test
Because things can take a bit longer the more addresses and events there
are, the initial proof retrieval attempt of the universe proof courier
can fail, causing it to go into a 30 second backoff.
Because we used a hard coded timeout of 30 seconds for the transfer to
complete, this sometimes failed the load test.
By making sure we use the same, configurable, timeout where relevant, we
make the send process more reliable.

In addition, the following lines should be added to the load test nodes
to shorten the duration of the backoff:

universerpccourier.initialbackoff=2s
universerpccourier.maxbackoff=30s