multi: refactor to split off tapsend package from tapscript
We have the following weird package dependency chain:
proof -> vm -> tapscript -> tappsbt

This makes things weird and cumbersome. The only parts in tapscript that
actually use the tappsbt stuff is send related.
So to break that dependency chain between the vm and tappsbt, we split
off all the send functionality into a new tapsend package.