address+tappsbt: add ProofDeliveryAddress to VOutput
When interactively signing a transfer that goes to a TAP address, the
participants might want to be able to check the proof courier delivery
URL that will be used to transfer the proof.
This information currently gets lost when converting a TAP address into
a vPSBT and is transported in a different data structure inside of the
freighter. For a flow that doesn't use the freighter (or just in the
very end), there currently is no way to specify the proof courier
address, which this commit now changes.
A follow-up commit will refactor the freighter to use this new field.