tappsbt: add ProofSuffix field to VOutput
We need a way to transport proofs that are complete in terms of MS-SMT
proofs committing to the correct asset but incomplete in terms of the
asset witness (in the ASSET_VERSION_V1 case).
The virtual transaction output is as good a place as any other. This
will also allow us to simplify some of the data structures around
passive assets, as currently we need to carry those proofs around in a
separate map.