tapdb+tapfreighter: refactor to use VPacket directly
This commit refactors the freighter in that we use the VPacket struct
directly for the passive assets. We now have all information available
directly in that struct and no longer need another wrapper.
With that looking more similar to other code, we can then also re-use
the updateAssetProofFile method which gets rid of some duplicated code.