multi: use anchor based output commitment everywhere
We had a mix of how we referred to output commitments before this
commit. In some places we had a slice and the index referred to the
output index of a virtual output within a virtual transaction. Then in
other places we used the anchor transaction output index as a map key.
To avoid mistakes of mixing the two types of indexes, we unify the code
to only ever use the anchor output index as the map key (and no more
slices with implicit indexing).