tapfreighter+tapsend: refactor CreateAnchorTx
In this commit we refactor the CreateAnchorTx in a way that will allow
us to have active and passive outputs that don't necessarily overlap in
their anchor output. Meaning that passive assets can be anchored to an
anchor output independent of any "active" transfer. To enable this, we
need to pass in _all_ virtual transactions that should be committed to
into the function.