tapfreighter: move proof suffix creation code
The ConvertToTransfer function shouldn't house any proof specific logic
and instead just focus on converting active and passive virtual
transactions into databse entries that represent transfers.
So we move the logic that creates the proof suffixes for the active
transfers to the same place where we do the same for the passive
transfers.

Since we're also going to make sure the passive assets have their own
anchor in the next commit, we can also remove the special logic around
the output type that only exists to carry passive assets.