multi: remove output commitments from AnchorTransaction
With the new RPC based flow we'll need to transport everything
within the AnchorTransaction struct over RPC. But commitments are
hard to serialize. And we only used the output commitments in
ConvertToTransfer to calculate the asset and merkle root of an output
anchor, we can just transport that information in the PSBT and remove
the need for carrying along the full output commitment.