tapfreighter: support multiple active transfers
This commit allows multiple active virtual transactions to be mapped to
the same transfer. We generally understand a "transfer" as everything
that happened on the asset level within a single bitcoin anchor
transaction.
Virtual transactions grouped into "active" transfers are movements that
the user explicitly initiated. These usually can be identified as
either creating a split (for partial amounts) or (for full value sends)
as sending to a script key.
Passive virtual transactions are state updates for all assets that were
committed to in the same anchor transaction as the active assets being
spent but remain under the daemon's control. These can be identified as
1-input-1-output virtual transactions that send to the same script key
as they were already committed at.