tapfreighter: give passive assets their own anchor
This is the meats and potatoes of the whole refactor. Everything up to
this point was mostly preparation.
In this commit we finally change the send logic and completely de-couple
the passive assets from the active ones.
Before this commit if we had passive assets but no active transfer that
sent an output back to ourselves, we created a fake "tombstone" output
with the tappsbt.TypePassiveAssetsOnly type, just to tell the database
storage logic where to anchor the passive assets.

This commit now attempts to find a suitable anchor transaction output
for attaching the passive assets to. If none can be found in the list of
active transfers, then a new anchor output is explicitly created for
them.