proof: return copy of mock proof
This fixes a unit test race condition where two goroutines would be
operating on the same proof because they both received the same
reference from the mock. This would not happen in production as the
proof archive always creates a new proof object.
So we fix the race by also creating a unique copy of the mocked proof
whenever it is requested.