itest: fix flake with universe sync
Fixes #807.

If we're syncing a node to a universe that is still in the process of
getting proofs pushed to, we sometimes receive leaves with a root that
doesn't match the root we fetched at the beginning of the sync process.
We should re-try the sync, which we'll do in a separate PR.
For now we just want to fix the flake, which we can do by waiting to
start a new node until the existing one has finished syncing.