tapfreighter: allow internal key to be unknown
If we want to have the ability to encumber the BTC level outputs by a
MuSig2 internal key and/or have a Tapscript sibling, we need to allow
for the internal key to be unknown to the wallet.
Otherwise we can't choose the NUMS pub key as the internal key if we
wanted, for example, to only allow a script spend on the BTC level.