multi: remove asset version from genesis
Because of the versions field being in the GROUP BY clause of the query,
we would get two distinct rows returned for the same asset ID if we had
asset UTXOs with different versions.
But because we use the asset ID as the map key of the data structure we
return, only one of those two entries would be returned, causing the
total amount to be incorrect.

But since the genesis doesn't really have a version in the database
anyway and we're just interested by the sum of asset balances by asset
ID, we remove the version field from the genesis entry which causes the
problem in the first place.