rpcserver+taprpc: return passive assets in FundVirtualPsbts
In the new RPC flow that's going to be added in the following commits,
the user will need to sign the passive assets manually.
We return the passive assets in the funding call so the user can
actually do that. For any existing flow these passive assets can just be
ignored as the signing still happens automatically under the hood.