tapsend+tapfreighter: move asset pruning
We no longer prune any tombstone or burn assets from the input
commitment, since we'll need them later when committing all assets
on-chain to validate that we can fully reconstruct the input TAP
commitment tree. So we keep the purgeable assets in the input commitment
and only filter them before creating the passive assets (which operates
on a copy anyway, so they stay in the original input commitment).