tapfreighter: prepare parcel for logging multiple vPackets
We'll want to support multiple virtual transactions as part of a
transfer within the porter at some point. For now we just prepare the
struct to hold a slice instead of a single packet.
And we add a new parcel type that will allow us to publish, log and
transfer proofs for a fully externally prepared and anchored
transaction.