tapfreighter: only select inputs belonging to output
This is a bug fix. Because a transfer can be the realization of multiple
virtual packets, not necessarily all of the same asset ID, we must
select the correct inputs when creating the proof suffix.
This was not really a problem until now because multiple virtual packets
weren't allowed, so this bug never showed itself.