address+tapgarden: skip completed events
With this commit we skip inspecting on-chain addresses for inbound
asset transfers that have already fully completed.
Previously, if there was an incomplete inbound transfer for a certain
block height, we would go through all transactions in blocks after that
height and notify for all inbound transfers found, even if they already
were fully complete.