tapfreighter: change order, send out event for send complete
Because the loop that sends out the notifications for each send
state terminates before reaching the SendStateComplete state, we
manually need to send out that event once we've sent out all proofs to
the receivers.
We also want to unify how and when we send out notifications, so we move
the event creation to _after_ executing an event.