tapgarden: update test asserts, remove ticker test
In this commit, we stop using the batch ticker in any minter unit tests.
This changed some test behavior, as finalizing via the ticker is
non-blocking. We update some assertions to read the DB directly instead
of querying the planter to maintain similar guarantees as before.
Finally, we remove a test section checking for batch cancellation after
finalize, which is not possible as finalize is blocking. This only
worked before because it used the batch ticker.