rfq: respond to an incoming sell request msg with a sell accept msg
In this commit we enhance the RFQ service such that the manager directs
an incoming (from a peer) asset sell quote request message to the
RFQ negotiator sub-service. The negotiator then populates the outgoing
messages channel with a sell request accept message. This message will
eventually be forwarded to the peer by the streamer RFQ sub-service.