rfq: register outgoing sell accept with order handler
This commit adds a new policy type to the order handler RFQ service. The
new policy type is specific to asset purchases (where the counterparty
has requested to sell an asset to our node). With this new type we can
check that a HTLC is valid in the context of an asset purchase
specifically.

In this commit we ensure that before an asset sell quote request accept
message is sent to our peer, the RFQ order handler registers a new
asset purchase policy.