rfq+rpc: rename to clarify that subject quotes were accepted by peers
In this commit, we rename methods, RPC endpoints, and variables to
clarify that the subject quotes are those that were requested by our
node and were accepted by our peers.

They are not quotes that our node has accepted. Quotes that our node has
accepted are stored as sale/purchase policies in the RFQ order handler.