itest: rename RFQ test to clarify payment init by asset buy request
In this commit we rename a test and update its doc to clarify that the
payment made as part of the test is initiated via an asset buy RFQ
request. This change is made in preparation for adding a future itest
where a payment is made following an asset sell RFQ request.