golangci: add lll linter with `new-from-rev` flag
This commit adds the lll linter to the golangci-lint configuration file.
This forces the code to be formatted with a maximum line length of 80
and a tab width of 8.

The `deadline` configuration is replaced with `timeout`. `deadline` is
deprecated since v1.20 and will be removed in v1.57.0.

The `new-from-rev` flag allows to run linters only on the files changed
since the given revision. Since we can assume that the code adheres to
linting standards before said revision there are no old errors that we
now ignore and this ensures that new code adheres to the additional
restrictions from `lll`.