server+tapchannelmsg: add custom channel data parser
We introduce a set of structs with JSON tags in order for us to be able
to serialize the channel custom data in a more human-readable way.
To avoid circular package dependencies, we need to split up the code for
the custom channel data parsing into two packages.
The main reason for this is that we use the JSON structs both in the rfq
and tapchannel packages.