proof: fix split commitment root issue with ownership proof
A recently introduced non-inflation check in the VM that takes the split
commitment root sum as the expected output sum of the VM pseudo
transaction caused ownership proofs to fail with split outputs.
This can be fixed trivially by not copying the split commitment root
when we have the change of a split asset.