rfq: validate incoming buy accept quote message
This commit includes new functionality to validate an incoming (sent
from an external peer) `BuyAccept` quote message. A `BuyAccept` quote
message is a response to the initiating node's request for quote
message.

The quote price is validated using a tolerance factor and a price
obtained from the initiating node's price oracle.