multi: add new declared_known flag to script key
We'll want to be able to detect that a script key is explicitly declared
as being known to a software that operates on the local node.
This is useful for example when the script key uses an internal key that
is not local (e.g. it cannot be derived by the local wallet, for example
a NUMS key), but the Tapscript spend paths are known to the software
operating on the tapd RPCs. So funds on such script keys should still
show up in the balances.