tapchannel: add initial version of AuxSweeper
In this commit, we implement the initial version of the AuxSweeper,
which implements both the `AuxSweeper` and `AuxResolver` methods defined
in lnd.

When we execute or detect a force close, we'll call `ResolveContract`
for each of the outputs. This'll insert any script keys we need into
the database, import the funding transaction proof for the responder,
and finally ship a transfer for the commitment transaction. This ensures
all proofs we need have been verified and written to disk.

For each contract, we can use the contract type to create a vPkt that is
able to sign for the contract input.

Once we're ready to sweep, the sweeper will call `DeriveSweepAddr` which
we'll use to batch sweep all the inputs (which contains the resolution
blob) into a single change address.

Once the sweeper is ready to broadcast, we call `NotifyBroadcast` which
allows us to complete the proofs for each of the outputs (now we know
the sweep transaction), then ship them off as a normal transfer.